pull context fetching logic into shared service
