+++ macos/Onit/UI/Panels/State/OnitPanelState+Input.swift
@@ -154,7 +154,7 @@ extension OnitPanelState {
         let windowApp = WindowHelpers.getWindowApp(pid: trackedWindow.pid)
         let appBundleUrl = windowApp?.bundleURL
         
-        AccessibilityNotificationsManager.shared.retrieveWindowContent(
+        ContextFetchingService.shared.retrieveWindowContent(
             state: self,
             trackedWindow: trackedWindow,
             customAppBundleUrl: appBundleUrl
